fix flow CI
