+++ flow-typed/package.json.js
@@ -1,0 +1,3 @@
+declare module '../package.json' {
+  declare var repository: string
+}
